A sample code for Huffman Decoding of time complexity: O(n + k^2)
Here is an algorithm of Huffman decoder with time complexity of
O(n+k^2), that takes as input a codeword dictionary containing the
codewords for k different symbols, and takes an encoded stream of n
bits, and decodes the sequence of encoded symbols.

Input specification: first line contains k, representing the number of
symbols, followed by k lines such that, i-th of these lines contains two
strings. The first string is a i-th symbol, and the second string is the
codeword used to encode that symbol. Following this is a line containing
n, representing the size of the encoded "bitstream". Following this is a
single line containing a single string of length n representing the
entire encoded "bitstream".

Output specification: a single line, containing the symbols decoded from
the encoded "bitstream".