Given n number of cards in random order, determine the number of moves required to sort them in O(n^2) time complexity.
Here is an O(n^2) algorithm that determines the minimum number of cards
that are needed to be moved in order to get the whole set of n cards in
sorted, ascending order. It is an application of longest Increasing
subsequence algorithm.

Input specification: the first line contains n and the next line
contains the current order of the cards from the first to the last,
separated by spaces.
Output specification: the output contains one line with the minimum
number of cards that need to be moved to achieve the sorted order.