Search for a path in the maze(matrix) in O(m*n) time complexity.
The following is the code to search if a path exist from a source to
destination in a given maze, in the form of a matrix of {0,1}. I have
implemented BFS search algorithm for the same. Given is a M*N matrix
with 0 representing a empty space/valid move while 1 representing a
wall, only single source (integer 2) and destination (integer 3).

Input specification: the first line contains m and n, separated by a
space. Then m lines follow. The i-th of these lines represents the i-th
row of the input array and it contains n numbers, each number is either
0, 1, 2 or 3 (and the numbers are separated by spaces).

Output: the output contains one line with the length of the shortest
path from 2 to 3, or the number -1 if no path from 2 to 3 exists.