Calculate the number of MST possible in the given graph in O(m log n).
Given a connected, undirected weighted graph, determine how many Minimum
Spanning Trees(MSTs) are possible.  The weight of the edges are not
always distinct, however it is guaranteed that for any cost there are
never more than two edges that have the same cost. Determine a O(m log
n) algorithm to determine the number of MSTs possible.
Here I have used my own data structure of Weighted graphs and the
Union-Find data structure and the Kruskal's algorithm is used to
determine the MST in the graph.

Input specification: The first line contains n and m (the number of
vertices and edges of the graph), separated by a space. The first line
is followed by m lines, each describe an edge. the vertices range from 0
to n-1.

Output specification: The output contains a single integer value,
indicating the number of minimum spanning trees that exist for the given
graph.