Detect if there exist a negative cycle in the given graph with time complexity of O(n^2).
I used the concept of strongly connected components of a graph to detect
the cycle and then determining the total cost it will take to traverse
once through the cycle, if the total cost is negative then there exist a
negative cycle in the given graph. The algorithm of BFS is used to
traverse the graph along with self defined data structure of Graph.

Input Specification: the first line contains n and m, the number of
vertices and edges. The vertices are labeled with numbers 1 to n. Then m
lines follow. The i-th line contains u, v, and w, specifying the i-th
edge that goes from vertex u to vertex v, of weight w.

Output specification: the first line contains the string "YES" if there
exists a negative-weight cycle, otherwise it contains the string "NO".